Move AsciiTree and FlatList to Output
